examples/kafka_docs_streaming/requirements.txt
firecrawl-py==1.9.0
kafka-python==2.0.2
pyyaml==6.0.1
python-dotenv==1.0.0
kubernetes==28.1.0